updated Paradiag object
