+++ asQ/paradiag_old.py
@@ -1,10 +1,10 @@
 import firedrake as fd
-from firedrake.petsc import PETSc
+from firedrake.petsc import flatten_parameters
+from firedrake.petsc import PETSc, OptionsManager
 from functools import partial
 from .profiling import memprofile
 
-from asQ.allatonce import AllAtOnceFunction, AllAtOnceForm, AllAtOnceSolver
-from asQ.allatonce.mixin import TimePartitionMixin
+from asQ.allatoncesystem import AllAtOnceSystem, JacobianMatrix
 from asQ.parallel_arrays import SharedArray
 
 __all__ = ['create_ensemble', 'paradiag']
@@ -38,39 +38,38 @@ def create_ensemble(time_partition, comm=fd.COMM_WORLD):
     return fd.Ensemble(comm, nspatial_domains)
 
 
-class paradiag(TimePartitionMixin):
+class paradiag(object):
     @memprofile
     def __init__(self, ensemble,
                  form_function, form_mass,
-                 ics, dt, theta,
-                 time_partition,
+                 w0, dt, theta,
+                 alpha, time_partition, bcs=[],
                  solver_parameters={},
-                 block_ctx={}, bcs=[],
-                 options_prefix="",
                  reference_state=None,
-                 function_alpha=None, jacobian_alpha=None,
-                 jacobian_function=None, jacobian_mass=None,
-                 pre_function_callback=None, post_function_callback=None,
-                 pre_jacobian_callback=None, post_jacobian_callback=None):
+                 linearised_function=None,
+                 linearised_mass=None,
+                 circ="",
+                 tol=1.0e-6, maxits=10,
+                 ctx={}, block_ctx={}, block_mat_type="aij"):
         """A class to implement paradiag timestepping.
 
         :arg ensemble: the ensemble communicator
         :arg form_function: a function that returns a form
-            on ics.function_space() providing f(w) for the ODE w_t + f(w) = 0.
+            on w0.function_space() providing f(w) for the ODE w_t + f(w) = 0.
         :arg form_mass: a function that returns a linear form on
-            ics.function_space providing the mass operator for the time derivative.
+            w0.function_space providing the mass operator for the time derivative.
         :arg linearised_function: a function that returns a form
-            on ics.function_space() which will be linearised to approximate
+            on w0.function_space() which will be linearised to approximate
             derivative(f(w), w) for the ODE w_t + f(w) = 0.
-        :arg linearised_mass: a function that returns a linear form on ics.function_space
+        :arg linearised_mass: a function that returns a linear form on w0.function_space
             providing which will be linearised to approximate the form_mass
-        :arg ics: a Function from ics.function_space containing the initial data.
+        :arg w0: a Function from w0.function_space containing the initial data.
         :arg dt: float, the timestep size.
         :arg theta: float, implicit timestepping parameter.
         :arg alpha: float, circulant matrix parameter.
         :arg time_partition: a list of integers, the number of timesteps
             assigned to each rank.
-        :arg bcs: a list of DirichletBC boundary conditions on ics.function_space.
+        :arg bcs: a list of DirichletBC boundary conditions on w0.function_space.
         :arg solver_parameters: options dictionary for nonlinear solver.
         :arg reference_state: a Function in W to use as a reference state
             e.g. in DiagFFTPC.
@@ -83,49 +82,90 @@ class paradiag(TimePartitionMixin):
         :arg tol: float, the tolerance for the relaxation method (if used)
         :arg maxits: integer, the maximum number of iterations for the
             relaxation method, if used.
+        :arg ctx: application context for solvers.
         :arg block_ctx: non-petsc context for solvers.
         :arg block_mat_type: set the type of the diagonal block systems.
             Default is aij.
         """
-        self.time_partition_setup(ensemble, time_partition)
 
-        # all-at-once function and form
+        self.aaos = AllAtOnceSystem(ensemble, time_partition,
+                                    dt, theta,
+                                    form_mass, form_function,
+                                    w0, bcs=bcs,
+                                    reference_state=reference_state,
+                                    linearised_function=linearised_function,
+                                    linearised_mass=linearised_mass,
+                                    circ=circ, alpha=alpha)
 
-        function_space = ics.function_space()
-        self.aaofunc = AllAtOnceFunction(ensemble, time_partition,
-                                         function_space)
-        self.aaofunc.set_all_fields(ics)
-
-        self.aaoform = AllAtOnceForm(self.aaofunc, dt, theta,
-                                     form_mass, form_function,
-                                     bcs=bcs, alpha=function_alpha)
-
-        # all-at-once jacobian
-        if jacobian_function is None:
-            jacobian_function = form_function
-        if jacobian_mass is None:
-            jacobian_mass = form_mass
-
-        self.jacobian_aaofunc = self.aaofunc.copy()
-
-        self.jacobian_form = AllAtOnceForm(self.jacobian_aaofunc, dt, theta,
-                                           jacobian_mass, jacobian_function,
-                                           bcs=bcs, alpha=jacobian_alpha)
-
+        self.ensemble = ensemble
+        self.layout = self.aaos.layout
+        self.time_partition = self.aaos.time_partition
+        self.time_rank = self.aaos.time_rank
+        self.nlocal_timesteps = self.aaos.nlocal_timesteps
+        self.ntimesteps = self.aaos.ntimesteps
+        self.alpha = self.aaos.alpha
+        self.tol = tol
+        self.maxits = maxits
+        self.circ = self.aaos.circ
+        self.ctx = ctx
         self.block_ctx = block_ctx
 
-        # all-at-once solver
+        # set up the PETSc Vecs (X for coeffs and F for residuals)
+        W = self.aaos.function_space
+
+        nlocal = self.nlocal_timesteps*W.node_set.size  # local times x local space
+        nglobal = self.ntimesteps*W.dim()  # global times x global space
+
+        self.X = PETSc.Vec().create(comm=ensemble.global_comm)
+        self.X.setSizes((nlocal, nglobal))
+        self.X.setFromOptions()
+        # copy initial data into the PETSc vec
+        with self.aaos.w_all.dat.vec_ro as v:
+            v.copy(self.X)
+        self.F = self.X.copy()
+
+        # sort out the appctx
         if "pc_python_type" in solver_parameters:
             if solver_parameters["pc_python_type"] == "asQ.DiagFFTPC":
                 appctx["paradiag"] = self
                 solver_parameters["diagfft_context"] = "asQ.paradiag.get_context"
+        self.solver_parameters = solver_parameters
+        self.flat_solver_parameters = flatten_parameters(solver_parameters)
 
-        self.solver = AllAtOnceSolver(self.aaoform, self.aaofunc,
-                                      solver_parameters=solver_parameters,
-                                      options_prefix=options_prefix,
-                                      jacobian_form=self.jacobian_form,
-                                      jacobian_reference_state=reference_state)
+        # set up the snes
+        self.snes = PETSc.SNES().create(comm=ensemble.global_comm)
+        self.opts = OptionsManager(self.flat_solver_parameters, '')
+        self.snes.setOptionsPrefix('')
 
+        # set up the residual function
+        def assemble_function(snes, X, Fvec):
+            self.aaos._assemble_function(snes, X, Fvec)
+
+        self.snes.setFunction(assemble_function, self.F)
+
+        # set up the Jacobian using the provided options
+        with self.opts.inserted_options():
+            self.jacobian = JacobianMatrix(self.aaos, self.snes)
+
+        Jacmat = PETSc.Mat().create(comm=ensemble.global_comm)
+        Jacmat.setType("python")
+        Jacmat.setSizes(((nlocal, nglobal), (nlocal, nglobal)))
+        Jacmat.setPythonContext(self.jacobian)
+        Jacmat.setUp()
+        self.Jacmat = Jacmat
+
+        def form_jacobian(snes, X, J, P):
+            # copy the snes state vector into self.X
+            X.copy(self.X)
+            self.aaos.update(X)
+            self.jacobian.update()
+            J.assemble()
+            P.assemble()
+
+        self.snes.setJacobian(form_jacobian, J=Jacmat, P=Jacmat)
+
+        # complete the snes setup
+        self.opts.set_from_options(self.snes)
         # iteration counts
         self.reset_diagnostics()
 
@@ -147,8 +187,8 @@ class paradiag(TimePartitionMixin):
 
         Must be called exactly once after each snes solve.
         """
-        self.linear_iterations += self.solver.snes.getLinearSolveIterations()
-        self.nonlinear_iterations += self.solver.snes.getIterationNumber()
+        self.linear_iterations += self.snes.getLinearSolveIterations()
+        self.nonlinear_iterations += self.snes.getIterationNumber()
         self.total_timesteps += sum(self.time_partition)
         self.total_windows += 1
 
@@ -179,11 +219,20 @@ class paradiag(TimePartitionMixin):
         for wndw in range(nwindows):
 
             preproc(self, wndw)
-            self.solver.solve()
+
+            with self.aaos.w_all.dat.vec_ro as v:
+                v.copy(self.X)
+
+            with self.opts.inserted_options():
+                self.snes.solve(None, self.X)
+
+            self.aaos.update(self.X)
+
             self._record_diagnostics()
+
             postproc(self, wndw)
 
-            converged_reason = self.solver.snes.getConvergedReason()
+            converged_reason = self.snes.getConvergedReason()
             is_linear = (
                 'snes_type' in self.flat_solver_parameters
                 and self.flat_solver_parameters['snes_type'] == 'ksponly'
