+++ examples/advection/periodic_dg_advection.py
@@ -13,12 +13,12 @@ parser = argparse.ArgumentParser(
     description='ParaDiag timestepping for scalar advection of a Gaussian bump in a periodic square with DG in space and implicit-theta in time. Based on the Firedrake DG advection example https://www.firedrakeproject.org/demos/DG_advection.py.html',
     formatter_class=argparse.ArgumentDefaultsHelpFormatter
 )
-parser.add_argument('--nx', type=int, default=64, help='Number of cells along each square side.')
+parser.add_argument('--nx', type=int, default=16, help='Number of cells along each square side.')
 parser.add_argument('--cfl', type=float, default=0.8, help='Convective CFL number.')
 parser.add_argument('--angle', type=float, default=pi/6, help='Angle of the convective velocity.')
 parser.add_argument('--degree', type=int, default=1, help='Degree of the scalar and velocity spaces.')
 parser.add_argument('--theta', type=float, default=0.5, help='Parameter for the implicit theta timestepping method.')
-parser.add_argument('--width', type=float, default=0.1, help='Width of the Gaussian bump.')
+parser.add_argument('--width', type=float, default=0.2, help='Width of the Gaussian bump.')
 parser.add_argument('--nwindows', type=int, default=1, help='Number of time-windows.')
 parser.add_argument('--nslices', type=int, default=2, help='Number of time-slices per time-window.')
 parser.add_argument('--slice_length', type=int, default=2, help='Number of timesteps per time-slice.')
@@ -95,7 +95,7 @@ def form_mass(q, phi):
 def form_function(q, phi):
     # upwind switch
     n = fd.FacetNormal(mesh)
-    un = 0.5*(fd.dot(u, n) + abs(fd.dot(u, n)))
+    un = fd.Constant(0.5)*(fd.dot(u, n) + abs(fd.dot(u, n)))
 
     # integration over element volume
     int_cell = q*fd.div(phi*u)*fd.dx
@@ -112,8 +112,8 @@ def form_function(q, phi):
 # The PETSc solver parameters used to solve the
 # blocks in step (b) of inverting the ParaDiag matrix.
 block_parameters = {
-    'ksp_type': 'gmres',
-    'pc_type': 'bjacobi',
+    'ksp_type': 'preonly',
+    'pc_type': 'lu',
 }
 
 # The PETSc solver parameters for solving the all-at-once system.
@@ -131,25 +131,24 @@ block_parameters = {
 #    'ksp_type': 'preonly'
 
 paradiag_parameters = {
+    'snes_type': 'ksponly',
     'snes': {
-        'linesearch_type': 'basic',
         'monitor': None,
         'converged_reason': None,
-        'rtol': 1e-10,
-        'atol': 1e-12,
-        'stol': 1e-12,
+        'rtol': 1e-8,
     },
     'mat_type': 'matfree',
-    'ksp_type': 'preonly',
+    'ksp_type': 'fgmres',
     'ksp': {
         'monitor': None,
         'converged_reason': None,
-        'rtol': 1e-10,
-        'atol': 1e-12,
-        'stol': 1e-12,
+        'rtol': 1e-8,
     },
     'pc_type': 'python',
-    'pc_python_type': 'asQ.DiagFFTPC'
+    'pc_python_type': 'asQ.ParaDiagPC',
+    'diagfft_alpha': args.alpha,
+    'diagfft_state': 'linear',
+    'aaos_jacobian_state': 'linear',
 }
 
 # We need to add a block solver parameters dictionary for each block.
@@ -164,18 +163,17 @@ for i in range(window_length):
 # Give everything to asQ to create the paradiag object.
 # the circ parameter determines where the alpha-circulant
 # approximation is introduced. None means only in the preconditioner.
-pdg = asQ.paradiag(ensemble=ensemble,
+pdg = asQ.Paradiag(ensemble=ensemble,
                    form_function=form_function,
                    form_mass=form_mass,
-                   w0=q0, dt=dt, theta=args.theta,
-                   alpha=args.alpha, time_partition=time_partition,
-                   solver_parameters=paradiag_parameters,
-                   circ=None)
+                   ics=q0, dt=dt, theta=args.theta,
+                   time_partition=time_partition,
+                   solver_parameters=paradiag_parameters)
 
 
 # This is a callback which will be called before pdg solves each time-window
 # We can use this to make the output a bit easier to read
-def window_preproc(pdg, wndw):
+def window_preproc(pdg, wndw, rhs):
     PETSc.Sys.Print('')
     PETSc.Sys.Print(f'### === --- Calculating time-window {wndw} --- === ###')
     PETSc.Sys.Print('')
@@ -195,12 +193,12 @@ if is_last_slice:
 
 # This is a callback which will be called after pdg solves each time-window
 # We can use this to save the last timestep of each window for plotting.
-def window_postproc(pdg, wndw):
+def window_postproc(pdg, wndw, rhs):
     if is_last_slice:
         # The aaos is the AllAtOnceSystem which represents the time-dependent problem.
         # get_field extracts one timestep of the window. -1 is again used to get the last
         # timestep and place it in qout.
-        pdg.aaos.get_field(-1, index_range='window', wout=qout)
+        pdg.aaofunc.get_field(-1, uout=qout)
         timeseries.append(qout.copy(deepcopy=True))
 
 
@@ -225,14 +223,15 @@ PETSc.Sys.Print(f'linear iterations: {pdg.linear_iterations}  |  iterations per 
 
 # Number of iterations needed for each block in step-(b), total and per block solve
 # The number of iterations for each block will usually be different because of the different eigenvalues
-PETSc.Sys.Print(f'block linear iterations: {pdg.block_iterations._data}  |  iterations per block solve: {pdg.block_iterations._data/pdg.linear_iterations}')
+block_iterations = pdg.solver.jacobian.pc.block_iterations
+PETSc.Sys.Print(f'block linear iterations: {block_iterations.data()}  |  iterations per block solve: {block_iterations.data()/pdg.linear_iterations}')
 
 # We can write these diagnostics to file, along with some other useful information.
 # Files written are: aaos_metrics.txt, block_metrics.txt, paradiag_setup.txt, solver_parameters.txt
 asQ.write_paradiag_metrics(pdg)
 
 # Make an animation from the snapshots we collected and save it to periodic.mp4.
-if is_last_slice:
+if is_last_slice and False:
 
     fn_plotter = fd.FunctionPlotter(mesh, num_sample_points=args.nsample)
